Decouple this crate from crates.io's database pool
The default for this crate will be r2d2. crates.io uses its own wrapper
around that to use a re-entrant mutex in order to share a connection in
tests, so we need to use a trait to represent a connection pool in order
to accomodate that.

It'd be great if we could provide some default configuration (pool size,
timeout), but since we expect the pool to be shared with the environment
that's not an option right now. In the future I might rejigger the API
so we give the connection pool to the jobs that need it, but I'll need
to think about how to make that ergonomic