Add some integration tests (currently failing)
I want to change the behavior of `run_all_pending_jobs` to not require
its own database connection. I was hoping to get some more tests in
place on that function, but it turns out that the signature doesn't even
do what I wanted (yet). I could make this test by adding a second
barrier to the test, and waiting for the first before making assertions,
but I'd rather not add that complexity to just remove it in the next
commit.