Change the behavior of `run_all_pending_jobs` to wait for jobs to run
Before this would do a query to see how many jobs could be retried, and
then loop that many times queueing them up, and immediately return. This
was problematic for a few reasons.

- It was hard to test. We needed to do explicit synchronization to make
  sure jobs were actually running (we're still doing that in tests that
  care about locking behavior)
- It never communicated when a catastrophic error has occurred. If the
  thread pool or connection pool become poisoned, we'd never find out.
  We'd just blindly enqueue jobs which would immediately fail
- It didn't determine whether jobs were actually locked or not.
- It required an extra connection in the pool for the count
- The runner never knew if jobs actually started or not. Ultimately if
  jobs aren't running, we want to kill the runner and restart it.

The new system works as follows:

- Create a new channel for the worker threads to communicate with the
  coordinating thread
- Queue as many jobs to run as there are available threads, or 1 if no
  threads are available and we've *never* queued a job
  - We need to queue at least 1 job to make sure we receive a message.
    Workers queued before the call to `run_all_pending_jobs` (e.g. they
    were run by the previous call) won't be able to communicate on our
    channel. We need at least one new worker to be executed so we find
    out if there are more jobs to run
- Listen for messages on our channel. Workers communicate if they got a
  job, there are no jobs left, or if an error occurred loading the job
  - The error cases here are only on *loading* a job, not running it.
    An error means either there's no database connection available, or
    we got an error loading the job from the DB. Both cases mean we
    probably need to restart the runner or change the configuration.
- Loop until we get an error message or a message that no jobs are left

This should ensure that our thread pool remains fully saturated, and
that we're returning any errors that occurred. We technically can't
guarantee that we return all errors, since job a could lock a job, job
b tries to lock a job gets none, reports no jobs left, and then job b
fails during deserialization.

I opted to go with the looser guarantee that errors are returned since
an error happening in deserialization is extermely unlikely, and if we
wait for all messages to be received before returning we are less likely
to keep the worker threads fully saturated.