Remove the need to register jobs with the runner
Instead of requiring that the runner call `.register` for every job it
wants to be able to run (which will result in a runtime error if any
queued job wasn't registered with the runner), we instead require that
jobs be registered when they are defined.

This is an improvement for a few reasons:

- Since a job has to be defined somewhere to be enqueued, it's easier to
  enforce that all defined jobs are registered if registration happens
  at the definition site
- Since `impl Job` is intended to eventually be generated, we can
  eventually get to the point where users never have to care about
  registration at all.
- The API we're leaning towards for generating job impls doesn't give a
  clear place to define a struct name, so we need *some* ergonomic way
  to register the jobs

This uses the `inventory` crate, which uses black magic and linker
attributes to define some functions which are run before main, which
populates a list of all jobs that were registered. When constructing the
registry, we then look for all jobs that have the appropriate env type
(in practice this will be all jobs, except for our test suite), and put
those in our `HashMap`.

Surprisingly, this didn't require any unsafe code. We could make this a
bit faster by passing the env as a `*const ()`, rather than using `&Any`
with a runtime check, but our performance is dominated by IO, so the
extra runtime check to ensure we don't blow up folks computers if a bug
slips in is worth it.