Actually mark that the correct number of messages are pending
I stupidly put += 1 which means every loop we will always queue at least
one job to be run if there are no idle threads. This is really awkward
to test since it has no outwardly visible side effects. The main
difference is that with the correct behavior, we will queue at most
`max_threads` jobs to run when there are no jobs to run, instead of
`max_threads * 2`. The goal is to avoid the pointless load on the DB
server.