Intoduce codegen for background jobs
This replaces manual implementations of `Job` with a custom proc macro
attribute. This attribute goes onto a function with the appropriate
signature, and will generate the job struct, job impl, and job
registration.

If the first argument is a reference, it's assumed to be the environment
type. Taking a reference for any other reason is invalid, since the
resulting struct has to be owned. Someone attempting to use a reference
for anything other than the env will get errors about missing lifetimes,
and if they try to add them they'll get an error that we don't accept
generic functions.

We will generate a function that returns the job struct. Example usage
looks like this:

    #[swirl::background_job]
    fn do_work(env: &MyEnv, arg1: String, arg2: String) -> Result<(), swirl::PerformError> {
        // arg1 and arg2 were serialized to the database, env comes from
        // the runner
    }

    fn do_work_later(conn: &PgConnection) -> Result<(), swirl::EnqueueError> {
        do_work("a".into(), "b".into()).enqueue(conn)
    }

The vast majority of the code is validating input, since we accept a
very small subset of valid Rust functions. The only thing that we could
theoretically accept that we don't currently are arguments with any
pattern other than an identifier.

While I don't expect folks to be referencing the generated job type, the
generated type will have the path `fn_name::Job`, so it should be easy
to document and for folks to find if it's ever needed.