Add a configurable timeout for jobs starting
This only affects the behavior of `run_all_pending_jobs`. This should be
set to well over the maximum time a job is expected to run. The
intention here is that if an unexpected fatal error occurs (connection
pool or thread pool are corrupted, jobs are hanging forever, etc), we
return an error from `run_all_pending_jobs` so the runner can be killed
and restarted.

This doesn't currently set a timeout on getting a connection from the
pool, that depends on https://github.com/sfackler/r2d2/pull/73 and will
be explicitly tested once we're giving the pool to jobs rather than
expecting it to be in the environment (there's some trickiness in place
to make sure we still run some jobs if the connection pool size is equal
to the thread pool size and only some jobs require a database
connection)