Ensure imports used in job bodies are considered used
If a trait is imported, but only used in the job body (which is in its
own module and only in scope because of `use super::*;`), the compiler
warns that the import is unused. This might be a bug in the compiler,
but we can work around it by moving the generated `Job` impl out of the
generated module.

Fixes #6