Don't require a serde dependency in consuming crates
Prior to serde_derive 1.0.90, there was no way for us to re-export
everything required for this derive to function without requiring that
all users of swirl have serde in their `Cargo.toml`. While it's probably
likely that anyone using this library is already using Serde, it's still
an arbitrary restriction.

Now we can re-export everything required from swirl directly. I've
specifically depended on `serde_derive` instead of `serde = { features
= ["derive"] }`, since we will only work with version 1.0.90 or later of
the derive crate, and serde does not have a strict version requirement
on the derive crate when the feature is enabled.