Don't panic when a job fails to load
This came up during use on crates.io. Right the only time we explicitly
panic a worker thread is when the transaction returns an error, which
happens if a job failed to load or if the update/delete after attempting
to run a job failed to execute.

Notably the job failing to load case can happen if the database is
read-only, which is something that crates.io has tests for. If the job
failed to load, we already informed the consumer since
`run_all_pending_jobs` will have returned an error. If we then proceed
to also panic the thread, `assert_no_failed_jobs` will fail, even if
there are no jobs in the DB to actually run.

I'm torn on whether we should panic or not in the case of a job failing
to update/delete. I've left it as a panic for now, since I think we
probably still want `assert_no_failed_jobs` to fail in this case, but we
should never be able to lock a job and fail to update/delete it unless
the database went down while the job was running, so it should never
really affect tests. In production as long as it shows up in logs we
don't really care. I'll probably change this to not panic when we add
more proper logging.

Fixes #7
Fixes #8