Integrate connection pool configuration into the runner
This moves the responsibility of constructing the database connection
pool into the runner's builder. This lets us reduce the amount of
boilerplate when no configuration is needed, and provide a reasonable
default pool size of 2 * the thread pool size (meaning if every job
requires a database connection, we can still utilize every thread).

With this change, there is no public API which gives a reference to the
runner's connection pool. The main reason to do this would be to put the
same pool on the job environment. The plan is to allow jobs to take a
connection as an argument to support this use case. This will be done in
a separate commit.