Re-enable injecting a fully constructed connection pool
crates.io needs this behavior in tests, where we want to make sure that
everything uses the same single connection which is inside a
transaction.

This commit brings back the ability to provide your own pool, which was
removed in #18. To accomplish this we had to move some things around in
funky ways. The end result is that the ability to give a DB url or a
partially configured pool is now only available for r2d2. To use
another pool, you just need to pass it in manually.